requirements.txt
absl-py==2.0.0
addict==2.4.0
aiofiles==22.1.0
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
airway==0.2.3
annotated-types==0.6.0
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv_latex_cleaner==1.0.5
asciimatics==1.14.0
ase==3.22.1
asttokens==2.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bibtexparser==1.4.1
black==23.11.0
bleach==6.1.0
blenderproc==2.6.2
bokeh==3.1.1
Brotli==1.1.0
cached-property==1.5.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cmapy==0.6.6
cmocean==3.0.3
colorcet==3.0.1
comm==0.1.4
conda-pack==0.7.1
ConfigArgParse==1.7
configparse==0.1.5
contourpy==1.1.1
cssselect2==0.7.0
cycler==0.12.1
Cython==0.29.36
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
docker-pycreds==0.4.0
einops==0.4.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.0
fake-bpy-module-latest==20240122
fastjsonschema==2.18.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
freetype-py==2.4.0
frozenlist==1.4.0
fsspec==2023.12.2
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.40
google-auth==2.23.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
hesiod==1.0.0
html5lib==1.1
huggingface-hub==0.20.1
idna==3.4
igraph==0.11.4
imageio==2.31.5
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
iopath==0.1.10
ipydatawidgets==4.3.5
ipykernel==6.25.2
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.8.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.4.0
jupyter_packaging==0.12.3
jupyter_server==2.8.0
jupyter_server_fileid==0.9.0
jupyter_server_proxy==4.1.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kaleido==0.2.1
keras==2.11.0
kiwisolver==1.4.5
latexcodec==3.0.0
libclang==16.0.6
libigl==2.5.0
lightning-utilities==0.9.0
livelossplot==0.5.5
loguru==0.7.2
lpips==0.1.4
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mesh2sdf==1.1.0
meshio==5.3.4
meshplot @ git+https://github.com/skoch9/meshplot.git@725e4a7926a5f10888f0edd1762fecf9db751c56
mistune==3.0.2
mmcv==1.5.3
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nibabel==5.2.0
ninja==1.11.1.1
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.24.4
oauthlib==3.2.2
odl==0.7.0
open3d==0.15.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
param==2.0.1
parso==0.8.3
pathspec==0.11.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip-autoremove==0.10.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
plotly==5.19.0
pluggy==1.4.0
plyfile==1.0.3
pooch==1.7.0
portalocker==2.8.2
progressbar==2.5
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.19.6
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py2cytoscape==0.7.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybtex==0.24.0
pycarus==0.7.6
pycparser==2.21
pyct==0.5.0
pydantic==2.5.3
pydantic_core==2.14.6
pyDeprecate==0.3.1
pydicom==2.4.4
pydot==2.0.0
pydotplus==2.0.2
pydyf==0.8.0
pyembree==0.2.11
pyfiglet==0.8.post1
pyglet==1.5.28
Pygments==2.16.1
pykdtree==1.3.7.post0
PyMCubes==0.1.4
PyOpenGL==3.1.0
pyparsing==3.1.1
pyphen==0.14.0
pyquaternion==0.9.9
pyrender==0.1.45
pysdf==0.1.9
pytest==8.0.1
python-dateutil==2.8.2
python-igraph==0.11.4
python-json-logger==2.0.7
pythreejs==2.4.2
pytorch-lightning==1.7.5
pytorch3d @ file:///local-scratch/localhome/asa409/projects/inr2vec/pytorch3d-0.7.2-cp38-cp38-linux_x86_64.whl#sha256=d417adde5650fbc0a07d7e9d3f285a86796fbc20dd81cbe100219741f1fcda6e
pytz==2023.3.post1
pyvista==0.42.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==24.0.1
rawpy==0.19.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
Rtree==1.1.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.8
safetensors==0.4.1
scikit-image==0.19.3
scikit-learn==1.3.1
scikit-video==1.1.11
scipy==1.10.1
scooby==0.8.0
Send2Trash==1.8.2
sentry-sdk==1.32.0
setproctitle==1.3.3
simpervisor==1.0.0
SimpleITK==2.3.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminado==0.17.1
texttable==1.7.0
tflearn==0.5.0
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.12
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.1
torch @ file:///local-scratch/localhome/asa409/projects/inr2vec/torch-1.12.0%2Bcu113-cp38-cp38-linux_x86_64.whl#sha256=af11b85dfc75224198fe3693701c9488a7b6fcf9ef0df4c5c16949fd3ca8b89f
torch-cluster @ file:///local-scratch/localhome/asa409/projects/inr2vec/torch_cluster-1.6.0%2Bpt112cu113-cp38-cp38-linux_x86_64.whl#sha256=b055579cdf86d1cb2920b98fed496c2a5bb5c286789d4986b200a1cdd21bc3f1
torch-fidelity==0.3.0
torch-scatter @ file:///local-scratch/localhome/asa409/projects/inr2vec/torch_scatter-2.1.0%2Bpt112cu113-cp38-cp38-linux_x86_64.whl#sha256=78a23a491d7b78513ba2a3c8ec8e20b2fc6bbc52096c4cccc1bb46db83754850
torch-sparse @ file:///local-scratch/localhome/asa409/projects/inr2vec/torch_sparse-0.6.16%2Bpt112cu113-cp38-cp38-linux_x86_64.whl#sha256=98880ed0db50bb57b814222015b64b73190250a66b0c7445841038a508520846
torch-spline-conv @ file:///local-scratch/localhome/asa409/projects/inr2vec/torch_spline_conv-1.2.1%2Bpt112cu113-cp38-cp38-linux_x86_64.whl#sha256=17ee0cac4a572c44ab8142ae8c855278c8293b099bfe65519d52541d783dd7c1
torch_geometric==2.4.0
torchmetrics==0.11.4
torchvision @ file:///local-scratch/localhome/asa409/projects/inr2vec/torchvision-0.13.0%2Bcu113-cp38-cp38-linux_x86_64.whl#sha256=2814ef684140524e11b42e091fd4bfd8a85cbf15f8287a3d3e1ba40495657ed7
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
traittypes==0.2.1
trame==3.2.7
trame-client==2.12.6
trame-server==2.12.0
trame-vtk==2.5.9
trame-vuetify==2.3.1
trimesh==3.23.5
typeguard==2.13.3
typer==0.9.0
types-python-dateutil==2.8.19.14
types-setuptools==57.4.18
typing_extensions==4.9.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
vtk==9.2.6
wandb==0.15.12
wcwidth==0.2.8
weasyprint==61.0
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.0
widgetsnbextension==4.0.9
wrapt==1.15.0
wslink==1.12.3
xrfeitoria==0.5.1
xrprimer==0.7.1
xyzservices==2023.10.1
y-py==0.6.2
yacs==0.1.8
yapf==0.40.2
yarl==1.9.2
ypy-websocket==0.8.4
zipp==3.17.0
zopfli==0.2.3
